+++ config/routes.rb
@@ -9,7 +9,7 @@ Rails.application.routes.draw do
 
   resources :users,
     :controller => 'clearance/users',
-    :only => [:create] do
+    :only => Clearance.configuration.user_actions do
       resource :password,
         :controller => 'clearance/passwords',
         :only => [:create, :edit, :update]
@@ -17,5 +17,8 @@ Rails.application.routes.draw do
 
   get '/sign_in' => 'clearance/sessions#new', :as => 'sign_in'
   delete '/sign_out' => 'clearance/sessions#destroy', :as => 'sign_out'
-  get '/sign_up' => 'clearance/users#new', :as => 'sign_up'
+
+  if Clearance.configuration.allow_sign_up?
+    get '/sign_up' => 'clearance/users#new', :as => 'sign_up'
+  end
 end